% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{dollarOHLCV}
\alias{dollarOHLCV}
\title{Dollars (or other monies) traded based OHLC bars}
\usage{
dollarOHLCV(ticks, target_volume)
}
\arguments{
\item{ticks}{Tick trade data.}

\item{target_volume}{The target volume for each bar.}
}
\value{
OHLCV bars of similar money traded.
}
\description{
Make OHLC bars containing at least a certain amount of money traded.
}
