% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{makeTimeOHLCV}
\alias{makeTimeOHLCV}
\title{Make time based OHLC bars}
\usage{
makeTimeOHLCV(ticks, align_period = 5L, align_by = "minutes")
}
\arguments{
\item{ticks}{The original trade data.}

\item{align_period}{Length of the time period to align by.}

\item{align_by}{Unit of the time period, can be:
\itemize{
\item seconds: one of \verb{'s', 'secs', 'seconds}
\item minutes: one of \verb{'m', 'mins', 'minutes}
\item hours: one of \verb{'h', 'hours}
\item days: one of \verb{'d', 'days'}
}}
}
\value{
Time based OHLC bars.
}
\description{
Make time based OHLC bars
}
\details{
Bars are aligned on multiples of the time period since midnight, that is a
tick happening at 01:06:41 will be in a bar starting on:
\itemize{
\item 01:06:40 if the bar length is 5 seconds
\item 01:06:30 if the bar length is 30 seconds
\item 01:06:00 if the bar length is 1 minute
\item 01:05:00 if the bar length is 5 minutes
\item 01:00:00 if the bar length is 1 hour
}
}
