% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{oneBarOHLCV}
\alias{oneBarOHLCV}
\title{Convert all ticks into one bar.}
\usage{
oneBarOHLCV(ticks)
}
\arguments{
\item{ticks}{The ticks to convert}
}
\value{
All ticks convert to one OHCLV bar
}
\description{
Convert all ticks into one bar.
}
