% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{tickOHLCV}
\alias{tickOHLCV}
\title{Tick based OHLC bars}
\usage{
tickOHLCV(ticks, num_ticks, prev_bar = NULL)
}
\arguments{
\item{ticks}{Input trades}

\item{num_ticks}{The number of ticks to group in.}

\item{prev_bar}{Bar previous to this set, will be extended with \code{trades} if
not complete.}
}
\description{
Tick based OHLC bars
}
