% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{timeOHLCV}
\alias{timeOHLCV}
\title{Make time based OHLC bars}
\usage{
timeOHLCV(
  ticks,
  align_period = 5L,
  align_by = "minutes",
  name = NULL,
  prev_bar = NULL,
  add_name = getOption("FancyBar.AddSymbolName")
)
}
\arguments{
\item{ticks}{The original trade data.}

\item{align_period}{Length of the time period to align by.}

\item{align_by}{Unit of the time period, can be:
\itemize{
\item seconds: one of \verb{'s', 'secs', 'seconds}
\item minutes: one of \verb{'m', 'mins', 'minutes}
\item hours: one of \verb{'h', 'hours}
\item days: one of \verb{'d', 'days'}
}}

\item{name}{Name of the symbol.}

\item{prev_bar}{Bar from a previous call to \code{timeOHLCV()} with an earlier
set of trades.}

\item{add_name}{Should the name of the symbol be added to the column names?}
}
\value{
Time based OHLC bars.
}
\description{
The traditional OHLC bars based on buckets of time.
}
\details{
Bars are aligned on multiples of the time period since midnight, that is a
tick happening at 01:06:41 will be in a bar starting on:
\itemize{
\item 01:06:40 if the bar length is 5 seconds
\item 01:06:30 if the bar length is 30 seconds
\item 01:06:00 if the bar length is 1 minute
\item 01:05:00 if the bar length is 5 minutes
\item 01:00:00 if the bar length is 1 hour
}

The \code{highfrequency} package uses the same logic but the timestamp of the bar
is the end of the bar period.
}
