% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{volumeOHLCV}
\alias{volumeOHLCV}
\title{Volume based OHLC bars}
\usage{
volumeOHLCV(ticks, target_volume, split_large_trades = FALSE)
}
\arguments{
\item{ticks}{Tick trade data.}

\item{target_volume}{The target volume for each bar.}

\item{split_large_trades}{Split trades larger than \code{target_volume} into 2 or
more bars.}
}
\value{
OHLCV bars of similar volume.
}
\description{
Make OHLC bars containing at least a certain amount of volume.
}
